Move wallet test utility to wallet/ directory
