+++ src/Makefile.test.include
@@ -106,7 +106,9 @@ BITCOIN_TESTS += \
 
 BITCOIN_TEST_SUITE += \
   wallet/test/wallet_test_fixture.cpp \
-  wallet/test/wallet_test_fixture.h
+  wallet/test/wallet_test_fixture.h \
+  wallet/test/test_tapyrus_wallet.cpp \
+  wallet/test/test_tapyrus_wallet.h
 endif
 
 test_test_tapyrus_SOURCES = $(BITCOIN_TEST_SUITE) $(BITCOIN_TESTS) $(JSON_TEST_FILES) $(RAW_TEST_FILES)
