+++ src/test/CMakeLists.txt
@@ -149,5 +149,6 @@ if(BUILD_TAPYRUS_WALLET)
 			../wallet/test/wallet_test_fixture.cpp
 			../wallet/test/wallet_tests.cpp
 			../wallet/test/create_transaction_tests.cpp
+			../wallet/test/test_tapyrus_wallet.cpp
 	)
 endif()
