+++ src/test/test_tapyrus.h
@@ -16,7 +16,6 @@
 #include <txdb.h>
 #include <txmempool.h>
 #include <consensus/consensus.h>
-#include <wallet/wallet.h>
 
 #include <memory>
 
@@ -134,27 +133,6 @@ struct TestMemPoolEntryHelper
     TestMemPoolEntryHelper &Height(unsigned int _height) { nHeight = _height; return *this; }
     TestMemPoolEntryHelper &SpendsCoinbase(bool _flag) { spendsCoinbase = _flag; return *this; }
     TestMemPoolEntryHelper &SigOpsCost(unsigned int _sigopsCost) { sigOpCost = _sigopsCost; return *this; }
-};
-
-class TestWalletSetup : public TestChainSetup {
-public:
-    TestWalletSetup(): TestChainSetup() {
-        initWallet();
-    }
-
-    ~TestWalletSetup() {
-        wallet.reset();
-    }
-
-    std::unique_ptr<CWallet> wallet;
-
-    bool ImportCoin(const CAmount amount);
-    bool IssueNonReissunableColoredCoin(const CAmount amount, ColorIdentifier& cid);
-    bool ProcessBlockAndScanForWalletTxns(const CTransactionRef tx);
-    bool AddToWalletAndMempool(const CTransactionRef tx);
-private:
-    void initWallet();
-    void Sign(std::vector<unsigned char>& vchSig, CKey& signKey, const CScript& scriptPubKey, int inIndex, CMutableTransaction& outTx, int outIndex);
 };
 
 #endif //TAPYRUS_TEST_TEST_TAPYRUS_H
